setRow函数遇到有bad（）问题

+++ query.cpp
@@ -6,7 +6,7 @@ public:
     size_t id;
     unsigned dist;
     ID_DIST_PAIR() : id(0), dist(0) {}
-    ID_DIST_PAIR(size_t k, unsigned distance) : id(k), dist(distance) {}
+    ID_DIST_PAIR(size_t k, float distance) : id(k), dist(distance) {}
     ID_DIST_PAIR(const ID_DIST_PAIR& pair) : id(pair.id), dist(pair.dist) {}
 };
 
@@ -41,7 +41,7 @@ public:
                 buffer[index] = pair;
             }
         } else {
-            for (int i = size; i != index; --i) {
+            for (size_t i = size; i != index; --i) {
                 if (i == capacity) {
                     continue;
                 }
